services/qtext-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
openai==1.3.7
google-cloud-vision==3.4.4
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
sqlalchemy==2.0.23
